refactor: migrate enums to constants for better organization
- Moved the `PaymentDetailsStatus` and `Gender` enums from `enums.ts` to their respective constants files for improved structure and maintainability.
- Updated imports in relevant components and schemas to reflect the new locations of the enums, ensuring consistent usage across the application.
