+++ src/lib/constants/bank-account.ts
@@ -54,3 +54,13 @@ export const BENEFICIARY_TYPES: CurrencyOption[] = [
   { value: "individual", label: "Individual" },
   { value: "business", label: "Business" },
 ];
+
+/**
+ * Enum representing the possible statuses of payment details
+ * Mirrors the database enum 'payment_details_status'
+ */
+export enum PaymentDetailsStatus {
+  PENDING = "pending",
+  APPROVED = "approved",
+  REJECTED = "rejected",
+}
